+++ changes.md
@@ -22,3 +22,10 @@
 * Introduced admonitions. They will look like regular blockquotes in normal rendered markdown, but will have special styling in PDF. (closes #2)
 * Tweaked styles for increased readability and broke out some hardcoded styles into the stylesheet.
 * Improved the command line interface to include support for folders, custom page sizes and custom margins.
+
+## 1.4.0
+
+* Added an imgsize argument for specifying the maximum allowed image size. (closes #4)
+* Added 'Important' as a warning-class keyword for admonitions. (closes #5)
+* Added 'linkify' to markdown-it to auto-convert urls to links. (closes #3)
+* Added a top margin to headings. (closes #6)